refactor: implement BigDecimal for finances, Law of Demeter in TareaHoja and robust State transitions

+++ src/main/java/com/workaudit/composite/ComponenteProyecto.java
@@ -1,6 +1,8 @@
 package com.workaudit.composite;
 
+import java.math.BigDecimal;
+
 public interface ComponenteProyecto {
-    double getCoste();
+    BigDecimal getCoste();
     String getName();
 }
